fixed stupid int vs. float error that caused rounding errors when moving
2003-08-30  Simon Budig  <simon@gimp.org>

	* app/tools/gimpvectortool.[ch]: fixed stupid int vs. float
	error that caused rounding errors when moving in a zoomed view.
	Fixed drawing artefact when connecting strokes did not succeed.

+++ app/tools/gimpvectortool.h
@@ -65,8 +65,8 @@ struct _GimpVectorTool
   GimpAnchorFeatureType restriction;    /* movement restriction           */
   gboolean              modifier_lock;  /* can we toggle the Shift key?   */
   GdkModifierType       saved_state;    /* modifier state at button_press */
-  gint                  last_x;         /* last x coordinate              */
-  gint                  last_y;         /* last y coordinate              */
+  gdouble               last_x;         /* last x coordinate              */
+  gdouble               last_y;         /* last y coordinate              */
                                                                    
   GimpAnchor        *cur_anchor;     /*  The current Anchor                */
   GimpStroke        *cur_stroke;     /*  The current Stroke                */
